Made some refactoring to the code
