+++ VGC/frontend/src/App.jsx
@@ -1,14 +1,16 @@
 import React from "react";
+import "./App.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Navbar from "./components/layout/Navbar";
-import Home from "./components/layout/Home";
-import Register from "./components/auth/Register";
-import Login from "./components/auth/Login";
+import Navbar from "./components/Navbar/Navbar";
+import HomePage from "./pages/HomePage/HomePage";
+import Register from "./components/Auth/Register";
+import Login from "./components/Auth/Login";
+import MainPage from "./pages/MainPage/MainPage";
 import NotFoundPage from "./pages/NotFoundPage/NotFoundPage";
 import SearchPage from "./pages/SearchPage/SearchPage";
 import GamePage from "./pages/GamePage/GamePage";
-import "./App.css";
-import "./components/auth/Auth.css";
+import "./components/Auth/Auth.css";
+import gameList from "./games.json";
 
 const App = () => {
   return (
@@ -16,11 +18,10 @@ const App = () => {
       <Navbar />
       <section className="containter">
         <Routes>
-          <Route path="/" element={<Home />} />
+          <Route path="/" element={<HomePage />} />
           <Route path="/register" element={<Register />} />
           <Route path="/login" element={<Login />} />
-          {/* need  different path */}
-          {/*<Route path="/" element={<MainPage gameList={gameList} />} /> */}
+          <Route path="/main" element={<MainPage gameList={gameList} />} />
           <Route path="/search" element={<SearchPage />} />
           <Route path="/game/:id" element={<GamePage />} />
           <Route path="*" element={<NotFoundPage />} />
