+++ VGC/frontend/src/components/auth/Auth.css
@@ -1,24 +1,13 @@
 @import url("https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;600;700&display=swap");
 
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: "MS Sans Serif", Geneva, sans-serif;
-}
-
-body {
-  background: #000;
-}
-
-body::before {
+.auth-background {
   content: "";
   position: absolute;
   left: 0;
   top: 0;
   width: 100%;
   height: 100%;
-  background: url("../../img/signup.jpg") no-repeat center center/cover;
+  background: url("../../assets/Signup.jpg") no-repeat center center/cover;
   background-position: center;
   z-index: -1;
 }
@@ -101,10 +90,6 @@ form button:hover {
   background: #000080;
 }
 
-.form-wrapper a {
-  text-decoration: none;
-}
-
 .form-wrapper a:hover {
   text-decoration: underline;
 }
@@ -126,9 +111,9 @@ form button:hover {
 .form-wrapper small a {
   color: #0071eb;
 }
-
+/* do we really need this? */
 @media (max-width: 740px) {
-  body::before {
+  .body::before {
     display: none;
   }
